scripts/requirements.txt
click
intelhex
pyserial
python-can